backend/export_requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
chromadb==0.4.18
numpy<2.0.0
pydantic==2.5.0
google-generativeai==0.3.2
google-genai
python-dotenv==1.0.0
psycopg2-binary==2.9.9
reportlab>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.0.0
